package.json
In this code, we are using the @celo/contractkit library to connect to the Celo network, interact with smart contracts, and optimize transactions. Here are the key steps:

Import the necessary modules: ContractKit and StableToken from the @celo/contractkit library.

Define the Celo network URL and create a new ContractKit instance using that URL.

Connect to the Celo network using kit.connection.connect().

Retrieve the user's wallet address using kit.web3.eth.getAccounts().

Get the Celo stable token contract using kit.contracts.getStableToken().

Approve the transfer of CELO tokens by calling the approve() function on the stable token contract.

Wait for the approval transaction to be mined using approvalTx.waitReceipt().

Transfer the CELO tokens to the recipient by calling the transfer() function on the stable token contract.

Wait for the transfer transaction to be mined using transferTx.waitReceipt().

Print a success message if the transactions are optimized successfully.